requirements.txt
Django>=1.10.5
django-cors-headers==2.0.1
django-oidc-provider==0.4.4
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
djoser==0.5.4
future==0.16.0
gunicorn==19.7.0
pycryptodomex==3.4.3
pyjwkest==1.3.0
PyJWT==1.4.2
requests==2.13.0
six==1.10.0
